Remove hardcoded storage devices on GCE
Storage devices in GCE are currently set either in the tfvars file or
hardcoded in TF/sls files. This is bad because device name in Linux are
handled by udev and the naming is dynamic (/dev/sdb could be /dev/sda
sometimes).

This commit removes all of these by using GCE functionalities to get the
correct device name.